+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_qr_scan","path":"/Users/eldiiaralmazbek/Desktop/cristian/flutter_qr_reader/","native_build":true,"dependencies":[]},{"name":"image_picker_ios","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_ios-0.8.6/","native_build":true,"dependencies":[]},{"name":"permission_handler_apple","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_apple-9.0.4/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.7/","native_build":true,"dependencies":[]},{"name":"flutter_qr_scan","path":"/Users/eldiiaralmazbek/Desktop/cristian/flutter_qr_reader/","native_build":true,"dependencies":[]},{"name":"image_picker_android","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_android-0.8.5+3/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"permission_handler_android","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_android-10.0.0/","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[{"name":"permission_handler_windows","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_windows-0.1.0/","native_build":true,"dependencies":[]}],"web":[{"name":"image_picker_for_web","path":"/Users/eldiiaralmazbek/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.8/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"flutter_qr_scan","dependencies":["image_picker"]},{"name":"image_picker","dependencies":["image_picker_android","image_picker_for_web","image_picker_ios"]},{"name":"image_picker_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"image_picker_for_web","dependencies":[]},{"name":"image_picker_ios","dependencies":[]},{"name":"permission_handler","dependencies":["permission_handler_android","permission_handler_apple","permission_handler_windows"]},{"name":"permission_handler_android","dependencies":[]},{"name":"permission_handler_apple","dependencies":[]},{"name":"permission_handler_windows","dependencies":[]}],"date_created":"2022-09-12 12:15:57.060694","version":"3.0.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_qr_scan","path":"/Users/eldiiar/Desktop/encointer/flutter_qr_reader/","native_build":true,"dependencies":[]},{"name":"image_picker_ios","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_ios-0.8.6/","native_build":true,"dependencies":[]},{"name":"permission_handler_apple","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_apple-9.0.4/","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_plugin_android_lifecycle","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_plugin_android_lifecycle-2.0.7/","native_build":true,"dependencies":[]},{"name":"flutter_qr_scan","path":"/Users/eldiiar/Desktop/encointer/flutter_qr_reader/","native_build":true,"dependencies":[]},{"name":"image_picker_android","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_android-0.8.5+3/","native_build":true,"dependencies":["flutter_plugin_android_lifecycle"]},{"name":"permission_handler_android","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_android-10.0.0/","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[{"name":"permission_handler_windows","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/permission_handler_windows-0.1.0/","native_build":true,"dependencies":[]}],"web":[{"name":"image_picker_for_web","path":"/Users/eldiiar/development/flutter/.pub-cache/hosted/pub.dartlang.org/image_picker_for_web-2.1.8/","dependencies":[]}]},"dependencyGraph":[{"name":"flutter_plugin_android_lifecycle","dependencies":[]},{"name":"flutter_qr_scan","dependencies":["image_picker"]},{"name":"image_picker","dependencies":["image_picker_android","image_picker_for_web","image_picker_ios"]},{"name":"image_picker_android","dependencies":["flutter_plugin_android_lifecycle"]},{"name":"image_picker_for_web","dependencies":[]},{"name":"image_picker_ios","dependencies":[]},{"name":"permission_handler","dependencies":["permission_handler_android","permission_handler_apple","permission_handler_windows"]},{"name":"permission_handler_android","dependencies":[]},{"name":"permission_handler_apple","dependencies":[]},{"name":"permission_handler_windows","dependencies":[]}],"date_created":"2022-09-13 10:53:47.313525","version":"3.3.1"}
