+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,13 +1,13 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/eldiiaralmazbek/development/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/eldiiaralmazbek/Desktop/cristian/flutter_qr_reader/example"
+export "FLUTTER_ROOT=/Users/eldiiar/development/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/eldiiar/Desktop/encointer/flutter_qr_reader/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
 export "DART_OBFUSCATION=false"
-export "TRACK_WIDGET_CREATION=false"
+export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"
 export "PACKAGE_CONFIG=.dart_tool/package_config.json"
