+++ example/lib/home.dart
@@ -3,10 +3,10 @@ import 'dart:io';
 
 import 'package:example/scan_view.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_qr_scan/flutter_qr_reader.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:permission_handler/permission_handler.dart';
 
-import 'package:flutter_qr_scan/flutter_qr_scan.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
