+++ example/lib/scan_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_qr_scan/flutter_qr_reader.dart';
 
-import 'package:flutter_qr_scan/qrcode_reader_view.dart';
 
 class ScanViewDemo extends StatefulWidget {
   const ScanViewDemo({Key? key}) : super(key: key);
