+++ lib/controller/controller.dart
@@ -1,0 +1,2 @@
+export 'flutter_qr_reader_image.dart';
+export 'qr_reader_view_controller.dart';
