+++ lib/flutter_qr_reader.dart
@@ -1,0 +1,5 @@
+library flutter_qr_scan;
+
+export 'controller/controller.dart';
+export 'paint/paint.dart';
+export 'views/views.dart';
