+++ lib/paint/paint.dart
@@ -1,0 +1,1 @@
+export 'qr_box_painter.dart';
