+++ lib/views/flutter_qr_scan.dart
@@ -1,32 +1,13 @@
-import 'dart:async';
-import 'dart:developer';
-import 'dart:io';
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
-class FlutterQrReader {
-  static const MethodChannel _channel = const MethodChannel('me.hetian.flutter_qr_reader');
+import 'package:flutter_qr_scan/flutter_qr_reader.dart';
 
-  static Future<String?> imgScan(File file) async {
-    if (file.existsSync() == false) {
-      return null;
-    }
-    try {
-      final rest = await _channel.invokeMethod(
-        "imgQrCode",
-        {"file": file.path},
-      );
-      return rest;
-    } catch (e) {
-      log(e.toString());
-      return null;
-    }
-  }
-}
 
 class QrReaderView extends StatefulWidget {
   final Function(QrReaderViewController)? callback;
@@ -54,6 +35,7 @@ class _QrReaderViewState extends State<QrReaderView> {
   Widget build(BuildContext context) {
     if (defaultTargetPlatform == TargetPlatform.android) {
       return AndroidView(
+        hitTestBehavior: PlatformViewHitTestBehavior.transparent,
         viewType: "me.hetian.flutter_qr_reader.reader_view",
         creationParams: {
           "width": (widget.width! * window.devicePixelRatio).floor(),
@@ -94,53 +76,5 @@ class _QrReaderViewState extends State<QrReaderView> {
   @override
   void dispose() {
     super.dispose();
-  }
-}
-
-typedef ReadChangeBack = void Function(String?, List<Offset>, String?);
-
-class QrReaderViewController {
-  QrReaderViewController(this.id) : _channel = MethodChannel('me.hetian.flutter_qr_reader.reader_view_$id') {
-    _channel.setMethodCallHandler(_handleMessages);
-  }
-  final int id;
-  final MethodChannel _channel;
-  late ReadChangeBack onQrBack;
-
-  Future _handleMessages(MethodCall call) async {
-    switch (call.method) {
-      case "onQRCodeRead":
-        final points = <Offset>[];
-        if (call.arguments.containsKey("points")) {
-          final pointsStrs = call.arguments["points"];
-          for (String point in pointsStrs) {
-            final a = point.split(",");
-            points.add(Offset(double.tryParse(a.first)!, double.tryParse(a.last)!));
-          }
-        }
-        String? rawData = '';
-        if (call.arguments.containsKey("rawData")) {
-          rawData = call.arguments["rawData"];
-        }
-
-        this.onQrBack(call.arguments["text"], points, rawData);
-        break;
-    }
-  }
-
-  // 打开手电筒
-  Future<bool?> setFlashlight() async {
-    return _channel.invokeMethod("flashlight");
-  }
-
-  // 开始扫码
-  Future<bool?> startCamera(ReadChangeBack onQrBack) async {
-    this.onQrBack = onQrBack;
-    return _channel.invokeMethod("startCamera");
-  }
-
-  // 结束扫码
-  Future<bool?> stopCamera() async {
-    return _channel.invokeMethod("stopCamera");
   }
 }
