+++ lib/views/qr_reader_view.dart
@@ -3,8 +3,9 @@ import 'dart:developer';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'flutter_qr_scan.dart';
 import 'package:image_picker/image_picker.dart';
+
+import 'package:flutter_qr_scan/flutter_qr_reader.dart';
 
 /// 使用前需已经获取相关权限
 /// Relevant privileges must be obtained before use
@@ -244,96 +245,4 @@ class QrcodeReaderViewState extends State<QrcodeReaderView> with TickerProviderS
     _clearAnimation();
     super.dispose();
   }
-}
-
-class QrScanBoxPainter extends CustomPainter {
-  const QrScanBoxPainter({
-    required this.animationValue,
-    required this.isForward,
-    this.boxLineColor,
-  });
-
-  final double animationValue;
-  final bool isForward;
-  final Color? boxLineColor;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final borderRadius = BorderRadius.all(Radius.circular(12)).toRRect(
-      Rect.fromLTWH(0, 0, size.width, size.height),
-    );
-    canvas.drawRRect(
-      borderRadius,
-      Paint()
-        ..color = Colors.white54
-        ..style = PaintingStyle.stroke
-        ..strokeWidth = 1,
-    );
-    final borderPaint = Paint()
-      ..color = Colors.white
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 2;
-    final path = Path();
-    // leftTop
-    path.moveTo(0, 50);
-    path.lineTo(0, 12);
-    path.quadraticBezierTo(0, 0, 12, 0);
-    path.lineTo(50, 0);
-    // rightTop
-    path.moveTo(size.width - 50, 0);
-    path.lineTo(size.width - 12, 0);
-    path.quadraticBezierTo(size.width, 0, size.width, 12);
-    path.lineTo(size.width, 50);
-    // rightBottom
-    path.moveTo(size.width, size.height - 50);
-    path.lineTo(size.width, size.height - 12);
-    path.quadraticBezierTo(size.width, size.height, size.width - 12, size.height);
-    path.lineTo(size.width - 50, size.height);
-    // leftBottom
-    path.moveTo(50, size.height);
-    path.lineTo(12, size.height);
-    path.quadraticBezierTo(0, size.height, 0, size.height - 12);
-    path.lineTo(0, size.height - 50);
-
-    canvas.drawPath(path, borderPaint);
-
-    canvas.clipRRect(BorderRadius.all(Radius.circular(12)).toRRect(Offset.zero & size));
-
-    // 绘制横向网格
-    final linePaint = Paint();
-    final lineSize = size.height * 0.45;
-    final leftPress = (size.height + lineSize) * animationValue - lineSize;
-    linePaint.style = PaintingStyle.stroke;
-    linePaint.shader = LinearGradient(
-      colors: [Colors.transparent, boxLineColor!],
-      begin: isForward ? Alignment.topCenter : Alignment(0.0, 2.0),
-      end: isForward ? Alignment(0.0, 0.5) : Alignment.topCenter,
-    ).createShader(Rect.fromLTWH(0, leftPress, size.width, lineSize));
-    for (int i = 0; i < size.height / 5; i++) {
-      canvas.drawLine(
-        Offset(
-          i * 5.0,
-          leftPress,
-        ),
-        Offset(i * 5.0, leftPress + lineSize),
-        linePaint,
-      );
-    }
-    for (int i = 0; i < lineSize / 5; i++) {
-      canvas.drawLine(
-        Offset(0, leftPress + i * 5.0),
-        Offset(
-          size.width,
-          leftPress + i * 5.0,
-        ),
-        linePaint,
-      );
-    }
-  }
-
-  @override
-  bool shouldRepaint(QrScanBoxPainter oldDelegate) => animationValue != oldDelegate.animationValue;
-
-  @override
-  bool shouldRebuildSemantics(QrScanBoxPainter oldDelegate) => animationValue != oldDelegate.animationValue;
 }
