+++ lib/views/views.dart
@@ -1,0 +1,2 @@
+export 'flutter_qr_scan.dart';
+export 'qr_reader_view.dart';
